b_fastapi/requirements.txt
genanki==0.12.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic
python-multipart==0.0.9
pymupdf
langchain
python-dotenv
appwrite==11.0.0

